deprecates the legacy smart camera web and remove reference from embed (#410)
* improve build process for webcomponents

* bump to 1.4.5 beta

* Update changelog

* copy readme and package files

* refactor components to src

* Add iife support
Update esbuild config

* Update changelog

* Fix duplicate attribution in selfie capture

* update route for require

* lint fix

* Fix import paths
Only push selected files to npm

* Only define components when not already defined

* Fix duplicate listener

* lint fix

* Gzip main files

* bump to version 2.0.0-beta

* fix tests
Update preview deployment to use web-components smart-camera-web

* fix lint

* fix iife rename

* update prod and staging workflows to use web-component smart-camera-web

* deprecates the legacy smart camera web and remove reference from embed

* Update .github/workflows/deploy.yml


* revert version bump

* deprecate version 1

* Fix tests
Fix smart-camera-web event listener in biometric-kyc

* Fix tests
Mimic old smart camera behaviour in enhanced docv

* remove deprecated event listners

* fix tests

* Remove .DS_Store files from repository

* fix id capture test

* Add test for hide-back-of-id

* make iife the default script

* revert changes

* revert gitignore

* revert changes

* update changelog

---------
