+++ packages/embed/cypress/tests/id-selection.cy.cjs
@@ -138,37 +138,53 @@ describe('No ID Selection', () => {
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-entry-screen')
+      .find('document-capture-instructions')
       .should('be.visible');
 
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-entry-screen #take-photo')
+      .find('document-capture-instructions#document-capture-instructions-front')
+      .shadow()
+      .find('#take-photo')
       .click();
 
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-camera-screen')
-      .should('be.visible');
-
-    cy.getIFrameBody()
-      .find('smart-camera-web')
-      .shadow()
-      .find('#capture-id-image')
-      .click();
-
-    cy.wait(2000);
-
-    cy.getIFrameBody()
-      .find('smart-camera-web')
-      .shadow()
-      .find('#id-camera-screen')
+      .find('document-capture-instructions#document-capture-instructions-front')
       .should('not.be.visible');
 
     cy.getIFrameBody()
       .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .should('be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .shadow()
+      .find('#capture-id-image')
+      .click();
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .should('not.be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture-review#front-of-document-capture-review')
+      .should('be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture-review#front-of-document-capture-review')
       .shadow()
       .find('#select-id-image')
       .click();
@@ -176,7 +192,7 @@ describe('No ID Selection', () => {
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-review-screen')
+      .find('document-capture-review#front-of-document-capture-review')
       .should('not.be.visible');
 
     cy.wait('@getUploadURL');
@@ -382,37 +398,53 @@ describe('Preselected Country', () => {
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-entry-screen')
+      .find('document-capture-instructions')
       .should('be.visible');
 
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-entry-screen #take-photo')
+      .find('document-capture-instructions#document-capture-instructions-front')
+      .shadow()
+      .find('#take-photo')
       .click();
 
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-camera-screen')
-      .should('be.visible');
-
-    cy.getIFrameBody()
-      .find('smart-camera-web')
-      .shadow()
-      .find('#capture-id-image')
-      .click();
-
-    cy.wait(2000);
-
-    cy.getIFrameBody()
-      .find('smart-camera-web')
-      .shadow()
-      .find('#id-camera-screen')
+      .find('document-capture-instructions#document-capture-instructions-front')
       .should('not.be.visible');
 
     cy.getIFrameBody()
       .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .should('be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .shadow()
+      .find('#capture-id-image')
+      .click();
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .should('not.be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture-review#front-of-document-capture-review')
+      .should('be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture-review#front-of-document-capture-review')
       .shadow()
       .find('#select-id-image')
       .click();
@@ -420,7 +452,7 @@ describe('Preselected Country', () => {
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-review-screen')
+      .find('document-capture-review#front-of-document-capture-review')
       .should('not.be.visible');
 
     cy.wait('@getUploadURL');
@@ -620,37 +652,53 @@ describe('Preselected Country and ID Type', () => {
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-entry-screen')
+      .find('document-capture-instructions')
       .should('be.visible');
 
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-entry-screen #take-photo')
+      .find('document-capture-instructions#document-capture-instructions-front')
+      .shadow()
+      .find('#take-photo')
       .click();
 
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-camera-screen')
-      .should('be.visible');
-
-    cy.getIFrameBody()
-      .find('smart-camera-web')
-      .shadow()
-      .find('#capture-id-image')
-      .click();
-
-    cy.wait(2000);
-
-    cy.getIFrameBody()
-      .find('smart-camera-web')
-      .shadow()
-      .find('#id-camera-screen')
+      .find('document-capture-instructions#document-capture-instructions-front')
       .should('not.be.visible');
 
     cy.getIFrameBody()
       .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .should('be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .shadow()
+      .find('#capture-id-image')
+      .click();
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture#document-capture-front')
+      .should('not.be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture-review#front-of-document-capture-review')
+      .should('be.visible');
+
+    cy.getIFrameBody()
+      .find('smart-camera-web')
+      .shadow()
+      .find('document-capture-review#front-of-document-capture-review')
       .shadow()
       .find('#select-id-image')
       .click();
@@ -658,7 +706,7 @@ describe('Preselected Country and ID Type', () => {
     cy.getIFrameBody()
       .find('smart-camera-web')
       .shadow()
-      .find('#id-review-screen')
+      .find('document-capture-review#front-of-document-capture-review')
       .should('not.be.visible');
 
     cy.wait('@getUploadURL');
