+++ packages/embed/package.json
@@ -30,7 +30,6 @@
     "@fingerprintjs/fingerprintjs": "^4.5.1",
     "@sentry/browser": "^8.33.0",
     "@sentry/esbuild-plugin": "^2.22.3",
-    "@smile_identity/smart-camera-web": "file:../smart-camera-web",
     "@smileid/web-components": "file:../web-components",
     "jszip": "^3.10.1",
     "ua-parser-js": "^2.0.2",
