+++ packages/embed/src/js/doc-verification.js
@@ -1,4 +1,6 @@
 import '@smileid/web-components/combobox';
+import '@smileid/web-components/smart-camera-web';
+
 import JSZip from 'jszip';
 import { version as sdkVersion } from '../../package.json';
 import { getMetadata } from './metadata';
@@ -104,11 +106,6 @@ import { getMetadata } from './metadata';
         event.data.includes('SmileIdentity::Configuration')
       ) {
         config = JSON.parse(event.data);
-        if (config.use_new_component) {
-          import('@smileid/web-components/smart-camera-web');
-        } else {
-          import('@smile_identity/smart-camera-web');
-        }
 
         LoadingScreen.querySelector('.credits').hidden =
           config.hide_attribution;
@@ -372,9 +369,6 @@ import { getMetadata } from './metadata';
           if (!documentCaptureConfig.has_back) {
             SmartCameraWeb.setAttribute('hide-back-of-id', true);
           }
-          if (documentCaptureConfig.has_back) {
-            SmartCameraWeb.setAttribute('capture-id', 'back');
-          }
 
           if (config.document_capture_modes) {
             SmartCameraWeb.setAttribute(
@@ -476,16 +470,6 @@ import { getMetadata } from './metadata';
   }
 
   SmartCameraWeb.addEventListener(
-    'imagesComputed',
-    (event) => {
-      images = event.detail.images;
-      setActiveScreen(UploadProgressScreen);
-      handleFormSubmit(event);
-    },
-    false,
-  );
-
-  SmartCameraWeb.addEventListener(
     'smart-camera-web.publish',
     (event) => {
       images = event.detail.images;
@@ -496,25 +480,9 @@ import { getMetadata } from './metadata';
   );
 
   SmartCameraWeb.addEventListener(
-    'backExit',
-    () => {
-      setActiveScreen(SelectIDType);
-    },
-    false,
-  );
-
-  SmartCameraWeb.addEventListener(
     'smart-camera-web.cancelled',
     () => {
       setActiveScreen(SelectIDType);
-    },
-    false,
-  );
-
-  SmartCameraWeb.addEventListener(
-    'close',
-    () => {
-      closeWindow(true);
     },
     false,
   );
